Check that paired reads have the same QNAME
This detects desynchronised input files, which occasionally happens
due to user error or system failure.

Checking the names just after printing them has no real performance
implications because the strings are already in cache.  (It might
be better to check while reading the input, but that would be more
complicated in the two-input-files case.)

Conflicts:
	bwamem_pair.c